pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-session-manager"
version = "0.1.0"
authors = [
    {name="Your Name", email="your@email.com"},
]
description = "AWS Session Manager for handling temporary credentials"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "boto3>=1.26.0",
    "python-dateutil>=2.8.2"
]

[project.urls]
Homepage = "https://github.com/yourusername/aws-session-manager"